+++ pkg/ent/dir/where.go
@@ -101,7 +101,7 @@ func Dir(v string) predicate.Dir {
 }
 
 // Level applies equality check predicate on the "level" field. It's identical to LevelEQ.
-func Level(v uint64) predicate.Dir {
+func Level(v int) predicate.Dir {
 	return predicate.Dir(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldLevel), v))
 	})
@@ -426,21 +426,21 @@ func DirContainsFold(v string) predicate.Dir {
 }
 
 // LevelEQ applies the EQ predicate on the "level" field.
-func LevelEQ(v uint64) predicate.Dir {
+func LevelEQ(v int) predicate.Dir {
 	return predicate.Dir(func(s *sql.Selector) {
 		s.Where(sql.EQ(s.C(FieldLevel), v))
 	})
 }
 
 // LevelNEQ applies the NEQ predicate on the "level" field.
-func LevelNEQ(v uint64) predicate.Dir {
+func LevelNEQ(v int) predicate.Dir {
 	return predicate.Dir(func(s *sql.Selector) {
 		s.Where(sql.NEQ(s.C(FieldLevel), v))
 	})
 }
 
 // LevelIn applies the In predicate on the "level" field.
-func LevelIn(vs ...uint64) predicate.Dir {
+func LevelIn(vs ...int) predicate.Dir {
 	v := make([]any, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -451,7 +451,7 @@ func LevelIn(vs ...uint64) predicate.Dir {
 }
 
 // LevelNotIn applies the NotIn predicate on the "level" field.
-func LevelNotIn(vs ...uint64) predicate.Dir {
+func LevelNotIn(vs ...int) predicate.Dir {
 	v := make([]any, len(vs))
 	for i := range v {
 		v[i] = vs[i]
@@ -462,28 +462,28 @@ func LevelNotIn(vs ...uint64) predicate.Dir {
 }
 
 // LevelGT applies the GT predicate on the "level" field.
-func LevelGT(v uint64) predicate.Dir {
+func LevelGT(v int) predicate.Dir {
 	return predicate.Dir(func(s *sql.Selector) {
 		s.Where(sql.GT(s.C(FieldLevel), v))
 	})
 }
 
 // LevelGTE applies the GTE predicate on the "level" field.
-func LevelGTE(v uint64) predicate.Dir {
+func LevelGTE(v int) predicate.Dir {
 	return predicate.Dir(func(s *sql.Selector) {
 		s.Where(sql.GTE(s.C(FieldLevel), v))
 	})
 }
 
 // LevelLT applies the LT predicate on the "level" field.
-func LevelLT(v uint64) predicate.Dir {
+func LevelLT(v int) predicate.Dir {
 	return predicate.Dir(func(s *sql.Selector) {
 		s.Where(sql.LT(s.C(FieldLevel), v))
 	})
 }
 
 // LevelLTE applies the LTE predicate on the "level" field.
-func LevelLTE(v uint64) predicate.Dir {
+func LevelLTE(v int) predicate.Dir {
 	return predicate.Dir(func(s *sql.Selector) {
 		s.Where(sql.LTE(s.C(FieldLevel), v))
 	})
